# src/main/resources/openapi3_0.yaml
openapi: 3.0.3
info:
  title: Training hub
  description: ''
  version: 1.0.11
tags:
  - name: Post
    description: ''
paths:
  /post:
    post:
      tags:
        - Post
      description: upload a post as a user
      operationId: uploadPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
      responses:
        '201':
          description: Created succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Bad response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid

    get:
      description: update a post as a user
      operationId: getPost
      tags:
        - Post
      summary: Get items using a Map object as a query parameter
      parameters:
        - name: filters
          in: query
          description: Key-value pairs to filter items
          schema:
            type: object
            additionalProperties:
              type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsList'
        '402':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Post Not Found

  /post/{postId}:
    put:
      description: update a post as a user
      operationId: updatePost
      tags:
        - Post
      parameters:
        - in: path
          name: postId
          required: true
          description: id of the post to be updated
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Not Found
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bad Request




components:
  schemas:
    PostRequest:
      type: object
      properties:
        title:
          type: string
          example: Spring boot
        description:
          type: string
          example: important in backend team
        tag:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        attachment:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    PostResponse:
      type: object
      properties:
        postId:
          type: integer
        title:
          type: string
          example: Springboot
        description:
          type: string
          example: important in backend team
        tag:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        attachment:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    Attachment:
      type: object
      properties:
        attachment:
          type: string
    Tag:
      type: object
      properties:
        tagName:
          type: string
    PostsList:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'